requirements.txt
streamlit==1.32.0
torch==2.2.0
diffusers==0.26.3
transformers==4.37.2
accelerate==0.27.2
pillow==10.2.0
huggingface-hub==0.20.3
safetensors==0.4.2
